Added menusystem, made LCDButtons a class
Breaking out the button functionality of the LCD.  It's a bit messy so
far, right now the buttons depend on the LCD, and they're using #DEFINEs
instead of actual enums.  But I want to get this working quickly :-)

Once the buttons/menu works, I'll need to add a thermistor module as
well, read the temperatures properly, and maybe an eeprom setting so I
can store good temperatures.